+++ artichoke-backend/src/extn/core/regexp/match_operator.rs
@@ -8,6 +8,7 @@ use crate::convert::{Convert, RustBackedValue, TryConvert};
 use crate::extn::core::matchdata::MatchData;
 use crate::extn::core::regexp::Regexp;
 use crate::sys;
+use crate::types::Int;
 use crate::value::Value;
 use crate::Artichoke;
 
@@ -99,7 +100,7 @@ pub fn method(interp: &Artichoke, args: Args, value: &Value) -> Result<Value, Er
             }
             (
                 matchdata,
-                Value::convert(interp, i64::try_from(match_pos.0).ok()),
+                Value::convert(interp, Int::try_from(match_pos.0).ok()),
             )
         } else {
             (matchdata, Value::convert(interp, None::<Value>))
